requirements.txt
langchain
langchain-ollama
langchain-community
langchain-text-splitters
langserve
fastapi
uvicorn
python-dotenv
faiss-cpu
requests
beautifulsoup4
lxml
email-validator